+++ demo/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>custom-crafter-api-demo</artifactId>
+    <artifactId>demo</artifactId>
     <groupId>online.aruka.custom_crafter</groupId>
     <packaging>jar</packaging>
