+++ demo/src/main/java/online/aruka/custom_crafter/demo/listener/CraftFailureListener.java
@@ -1,8 +1,8 @@
-package online.aruka.custom_crafter.listener;
+package online.aruka.custom_crafter.demo.listener;
 
 import io.github.sakaki_aruka.customcrafter.event.failure.CraftInputInterruptEvent;
 import io.github.sakaki_aruka.customcrafter.event.failure.PreventDoubleCraftEvent;
-import online.aruka.custom_crafter.Demo;
+import online.aruka.custom_crafter.demo.Demo;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
