+++ demo/src/main/kotlin/online/aruka/custom_crafter/demo/listener/CustomCrafterAPIPropertiesChangeListener.kt
@@ -1,4 +1,4 @@
-package online.aruka.custom_crafter.listener
+package online.aruka.custom_crafter.demo.listener
 
 import io.github.sakaki_aruka.customcrafter.event.CustomCrafterAPIPropertiesChangeEvent
 import io.github.sakaki_aruka.customcrafter.util.Converter.toComponent
