+++ demo/src/main/kotlin/online/aruka/custom_crafter/demo/listener/CustomItemRegisteredListener.kt
@@ -1,7 +1,7 @@
-package online.aruka.custom_crafter.listener
+package online.aruka.custom_crafter.demo.listener
 
 import io.github.sakaki_aruka.customcrafter.event.RegisterCustomRecipeEvent
-import online.aruka.custom_crafter.Demo
+import online.aruka.custom_crafter.demo.Demo
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
